test(toast): migrate trigger
